+++ server/players/Player.js
@@ -8,6 +8,9 @@ export class Player {
     this.x = x;
     this.y = y;
     this.direction = 'down';
+    this.facing = 'down';
+    this.aimX = x;
+    this.aimY = y;
     this.moving = false;
   }
 
@@ -19,6 +22,9 @@ export class Player {
       x: this.x,
       y: this.y,
       direction: this.direction,
+      facing: this.facing,
+      aimX: this.aimX,
+      aimY: this.aimY,
       moving: this.moving,
     };
   }